# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Commerce API
  description: "REST API para gestao de operacoes de comercio: catalogo, carrinho,\
    \ pedidos, pagamentos e relatorios."
  contact:
    name: Equipe Commerce API
    url: https://empresa.com/commerce-api
    email: commerce-api@empresa.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
...
servers:
- url: http://localhost:8080
  description: Ambiente local
- url: https://api-dev.empresa.com
  description: Ambiente de desenvolvimento
security:
- bearerAuth: []
...
paths:
...
  /me/addresses/{id}:
    put:
      tags:
      - Área do Cliente - Endereços
      operationId: updateAddress
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AddressResponse"
    delete:
      tags:
      - Área do Cliente - Endereços
      operationId: deleteAddress
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /reports/sales:
    post:
      tags:
      - Administração - Relatórios de Vendas
      operationId: summary
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesReportRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SalesSummaryResponse"
...
  /me/orders/checkout:
    post:
      tags:
      - Área do Cliente - Pedidos
      operationId: checkout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /me/cart/items:
    post:
      tags:
      - Área do Cliente - Carrinho
      operationId: addItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCartItemRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /me/cart/coupon:
    post:
      tags:
      - Área do Cliente - Carrinho
      operationId: applyCoupon
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyCouponRequest"
        required: true
      responses:
        "200":
          description: OK
  /me/addresses:
    get:
      tags:
      - Área do Cliente - Endereços
      operationId: myAddresses
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AddressResponse"
    post:
      tags:
      - Área do Cliente - Endereços
      operationId: saveAddress
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AddressResponse"
  /manual-sales:
    post:
      tags:
      - Administração - Venda Manual
      operationId: createSale
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualSaleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
  /financial/summary:
    post:
      tags:
      - Administração - Financeiro
      operationId: summary_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialSummaryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FinancialSummaryResponse"
...
  /auth/reset-password:
    post:
      tags:
      - Público - Autenticação
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
  /auth/login:
    post:
      tags:
      - Público - Autenticação
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/forgot-password:
    post:
      tags:
      - Público - Autenticação
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
...
  /affiliates/me/summary:
    post:
      tags:
      - Área do Cliente - Afiliado
      operationId: mySummary
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffiliateReportRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AffiliateSummaryResponse"
...
  /me/password:
    patch:
      tags:
      - Área do Cliente - Minha Conta
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "200":
          description: OK
  /me/email:
    patch:
      tags:
      - Área do Cliente - Minha Conta
      operationId: changeEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeEmailRequest"
        required: true
      responses:
        "200":
          description: OK
  /me/cart/items/{itemId}:
    delete:
      tags:
      - Área do Cliente - Carrinho
      operationId: removeItem
      parameters:
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - Área do Cliente - Carrinho
      operationId: updateItemQuantity
      parameters:
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
...
components:
  schemas:
...
    CreateAddressRequest:
      type: object
      properties:
        street:
          type: string
          minLength: 1
        number:
          type: string
          minLength: 1
        complement:
          type: string
        neighborhood:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        zipCode:
          type: string
          minLength: 1
        primaryAddress:
          type: boolean
      required:
      - city
      - neighborhood
      - number
      - primaryAddress
      - state
      - street
      - zipCode
    AddressResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        street:
          type: string
        number:
          type: string
        complement:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        primaryAddress:
          type: boolean
        userId:
          type: integer
          format: int64
        userName:
          type: string
...
    SalesReportRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        saleChannel:
          type: string
          enum:
          - ONLINE
          - MANUAL
    SalesSummaryResponse:
      type: object
      properties:
        totalSales:
          type: integer
          format: int64
        onlineSales:
          type: integer
          format: int64
        manualSales:
          type: integer
          format: int64
        totalRevenue:
          type: number
        onlineRevenue:
          type: number
        manualRevenue:
          type: number
        averageTicket:
          type: number
...
    CreateOrderRequest:
      type: object
      properties:
        addressId:
          type: integer
          format: int64
      required:
      - addressId
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productName:
          type: string
        productPrice:
          type: number
        quantity:
          type: integer
          format: int32
        subtotal:
          type: number
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        addressId:
          type: integer
          format: int64
        total:
          type: number
        status:
          type: string
          enum:
          - PENDENTE
          - PAGO
          - ENVIADO
          - ENTREGUE
          - CANCELADO
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
    CreateCartItemRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
      required:
      - productId
      - quantity
    CartItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        productPrice:
          type: number
        quantity:
          type: integer
          format: int32
    CartResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        total:
          type: number
        couponCode:
          type: string
        discountApplied:
          type: number
        totalWithDiscount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemResponse"
    ApplyCouponRequest:
      type: object
      properties:
        couponCode:
          type: string
    ManualSaleItemRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
    ManualSaleRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ManualSaleItemRequest"
        notes:
          type: string
...
    FinancialSummaryRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        saleChannel:
          type: string
          enum:
          - ONLINE
          - MANUAL
        expenseCategory:
          type: string
          enum:
          - COMPRA_PRODUTOS
          - MARKETING
          - EMBALAGENS
          - INFRAESTRUTURA
          - FRETE
          - IMPOSTOS
          - OUTROS
        includeAffiliateCommissions:
          type: boolean
    FinancialSummaryResponse:
      type: object
      properties:
        totalRevenue:
          type: number
        onlineRevenue:
          type: number
        manualRevenue:
          type: number
        totalExpenses:
          type: number
        affiliateCommissions:
          type: number
        resultadoBruto:
          type: number
        resultadoLiquido:
          type: number
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 1
      required:
      - newPassword
      - token
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - email
      - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
      required:
      - email
...
    AffiliateReportRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    AffiliateSummaryResponse:
      type: object
      properties:
        totalSales:
          type: number
        totalCommission:
          type: number
        pendingCommission:
          type: number
        paidCommission:
          type: number
...
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 1
      required:
      - currentPassword
      - newPassword
    ChangeEmailRequest:
      type: object
      properties:
        currentEmail:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
        newEmail:
          type: string
          format: email
          minLength: 1
      required:
      - currentEmail
      - newEmail
      - password
    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: integer
          format: int32
      required:
      - quantity
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT